move the function into a companion object instead of inside the trait
